APIs consertadas e APIS Usuario, Comunidades e Posts testadas no Postman

+++ java/src/main/java/com/backendpi/backend/controller/PostSalvoController.java
@@ -27,7 +27,7 @@ public class PostSalvoController {
     }
 
     @DeleteMapping("/{id}")
-    public void deletar(@PathVariable Integer id) {
+    public void deletar(@PathVariable Long id) {
         service.deletar(id);
     }
 }
